refactor: consolidate all JNI native builds into a single shared library managed by the filament module and introduce Java module descriptors.
