+++ java/filamat/build.gradle.kts
@@ -3,7 +3,13 @@ plugins {
     id("filament-publish")
 }
 
-apply(from = rootProject.file("java/java-cmake.gradle.kts"))
+// Native code lives in the combined libfilament-jni built by :java:filament.
+// This module is pure-Java glue; the dylib reaches it transitively at runtime.
+
+configure<JavaPluginExtension> {
+    sourceCompatibility = JavaVersion.VERSION_17
+    targetCompatibility = JavaVersion.VERSION_17
+}
 
 dependencies {
     implementation(project(":java:filament"))
