+++ java/filament-utils/src/main/java/module-info.java
@@ -1,0 +1,7 @@
+module io.github.erkko68.filament.utils.jni {
+    exports io.github.erkko68.filament.utils.jni;
+
+    requires transitive io.github.erkko68.filament.jni;
+    requires kotlin.stdlib;
+    requires static org.jetbrains.annotations;
+}
