+++ java/gltfio/src/main/java/io/github/erkko68/filament/gltfio/jni/Gltfio.java
@@ -1,17 +1,15 @@
 package io.github.erkko68.filament.gltfio.jni;
 
 import io.github.erkko68.filament.jni.Filament;
-import io.github.erkko68.filament.jni.internal.NativeLoader;
 
 public class Gltfio {
     private static boolean sInitialized = false;
 
     public static synchronized void init() {
         if (sInitialized) return;
+        // All JNI symbols (filament, filamat, gltfio, filament-utils) live in a single
+        // combined dylib loaded by Filament.init().
         Filament.init();
-        NativeLoader.load("gltfio-jni");
         sInitialized = true;
     }
 }
-
-
