+++ java/gltfio/src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module io.github.erkko68.filament.gltfio.jni {
+    exports io.github.erkko68.filament.gltfio.jni;
+
+    requires transitive io.github.erkko68.filament.jni;
+    requires static org.jetbrains.annotations;
+}
